refactor: move useApp hook into hooks folder
