+++ src/renderer/views/LoadingView.tsx
@@ -3,7 +3,7 @@ import { jsx } from "@emotion/react";
 import React from "react";
 
 import { LoadingScreen } from "@/components/LoadingScreen";
-import { useAppStore } from "@/store/app";
+import { useAppStore } from "@/lib/hooks/useApp";
 import { withSlippiBackground } from "@/styles/withSlippiBackground";
 
 export const LoadingView: React.FC = () => {
